Move BreadCrumbButton to public api
Moves the static BreadCrumbButton inner class from the inaccessible
impl.org.controlsfx.skin.BreadCrumbBarSkin class to the publicly
accessible org.controlsfx.controls.BreadCrumbBar class.
This allows users to customize the appearance of the BreadCrumbBar via
its setCrumbFactory method. While it is possible to call this method
with a generic Button, the internal layouting of the skin relies on
BreadCrumbButton to layout the buttons in an overlapping way.

This implements the solution outlined by @abhinayagarwal in
controlsfx/controlsfx#1109

solves controlsfx/controlsfx#1109
